requirements.txt
agno
ipykernel
python-dotenv
groq
openai
tavily-python
sqlalchemy
fastapi
uvicorn